requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
cycler==0.11.0
dill==0.3.4
flatbuffers==2.0
fonttools==4.28.5
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.1
networkx==2.8.2
numpy==1.22.1
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.0
promise==2.3
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
PyWavelets==1.3.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.2
scipy==1.8.1
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow-metadata==1.5.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.5.4
tqdm==4.62.3
typing==3.7.4.3
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
